@@ -7,7 +7,7 @@ import com.codedrills.model.recommendation.PracticeRecommendations;
 import com.codedrills.model.recommendation.RecommendationContext;
 import com.codedrills.service.recommenders.RecommenderService;
 import com.codedrills.util.HandleHelper;
-import org.apache.log4j.Logger;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -16,9 +16,8 @@ import java.util.*;
 import static com.codedrills.util.HandleHelper.canonicalHandles;
 
 @Service
+@Slf4j
 public class ProfileService {
-  private static Logger logger = Logger.getLogger(ProfileService.class);
-
   @Autowired
   AnalysisService analysisService;
   @Autowired
@@ -28,14 +27,14 @@ public class ProfileService {
 
     List<Handle> handles = HandleHelper.validateAndSplit(handlesQuery);
 
-    logger.info(String.format("Analysis started for %s", handlesQuery));
+    log.info(String.format("Analysis started for %s", handlesQuery));
     AnalysisContext analysisContext = analysisService.analyzeUser(handles);
-    logger.info(String.format("Analysis completed for %s", handlesQuery));
+    log.info(String.format("Analysis completed for %s", handlesQuery));
 
-    logger.info(String.format("Recommendations started for %s", handlesQuery));
+    log.info(String.format("Recommendations started for %s", handlesQuery));
     RecommendationContext recommendationContext = recommenderService.recommendations(analysisContext);
     PracticeRecommendations practiceRecommendations = new PracticeRecommendations(recommendationContext.getRecommendations());
-    logger.info(String.format("Recommendations completed for %s", handlesQuery));
+    log.info(String.format("Recommendations completed for %s", handlesQuery));
 
     return new AnalysisResult(canonicalHandles(handles), analysisContext.getUserStats(), practiceRecommendations);
   }
